Replace our hand-rolled BytesReader with std::io::Read and byteorder
Turns out plain old &[u8] provides std::io::Read behavior, which gives
us the same cursor-like behavior as BytesReader. The byteorder crate
provides helpers for reading big- and little-endian integers very
similar to the ones we had written. The one missing piece was the
ability to slice off a series of bytes without allocating; that really
does depend on the underlying type being a slice, or at least having
memory that's not going to go away as we read. That's been moved to a
new trait that's just implemented for slices.