Remember closed sockets for several ticks instead of just one
Previously, the epoll-based server would close sockets for unused
connections (those that had not had an ICE binding request in 60s) on
one tick, and then clear the entry for that IP address on the next
tick. This was to avoid the following scenario:

1. UDP handler produces packets for socket X.
2. The UDP handler is pre-empted.
3. The tick handler runs and removes socket X.
4. The UDP handler resumes and tries to send to socket X.
5. The UDP handler thinks it needs to make a new connection.

However, we're still seeing a slow leak of file descriptors, and in
fact it's accelerated with the new pacing support. One possibility is
that *two* ticks happen in step 3, something that was fairly unlikely
before but much more possible now that there's a pacer for outgoing
packets running on its own thread pool.

To fix this, closing a socket now leaves behind the timestamp when the
socket was closed, so that it can be cleared out several ticks later
instead of just on the next tick. This still isn't perfect, but if
this is our primary cause of leaks, it should help quite a bit.