Throttle keyframe requests on a per-sender basis
This was clearly intended in the original code, but not correctly
implemented. With this we should better handle keyframe requests that
come in just after sending a previous request (but before the keyframe
is delivered), at the cost of potentially delaying some *legitimate*
requests if one comes in right *after* a keyframe was delivered.