Add 5% jitter to the cleaner interval
Empirically, two frontend instances started at "the same time" end up
trying to clean the same tables at the same time as well, which
results in doing twice as much work as necessary. Add some jitter to
the interval between cleaning so that the multiple instances get out
of sync after a while.