googcc: Only honor the latest Request
One congestion control request is sent per tick, but if a particular
client is alone in a call, there won't be any packets being sent to
that client, which means there isn't any information to generate a
congestion control update. In this case requests were just piling up
in an unbounded queue. Oops.

Fix this by only keeping the latest request around and vending that
when polled. (We keep the stream implementation because it makes
testing easier.)