backend: Remove all_demux_ids list from SfuToDevice messages for now
RingRTC doesn't currently use it for anything, and the addition of
pending clients using the same DeviceJoinedOrLeft makes the situation
a little more complicated anyway. For now, save the (small) cost of
this list by leaving it out, except for in tests where it's used to
check the current state of the call.