frontend: Simplify UpsertableItem by omitting primary keys from serde
For CallRecord this means marking `room_id` as `skip_serializing`, but
for CallLinkState the room ID is always passed separately anyway, so
we can omit it from the struct altogether.

Any caller already has it; it's used to query for the state. Doing
this makes UpsertableItem simpler; we can just assume the item in
question does not contain any primary keys.