Lock and synchronization work
Simplify making changes to the calling service by moving the responsibility for lock management into the objects that contain the data that needs to be protected.

Reduce lock contention by converting Mutexes to RwLocks where appropriate and by carefully introducing more locks so that a more fine grained locking strategy can be employed.